+++ features/ajna/common/components/AjnaProductCardsWrapper.tsx
@@ -6,7 +6,7 @@ interface AjnaProductCardsWrapperProps {
 }
 
 export const AjnaProductCardsWrapper = ({ children }: AjnaProductCardsWrapperProps) => (
-  <ProductCardsWrapper sx={{ mt: ['9px', '48px'], gap: ['88px', 3, 3] }}>
+  <ProductCardsWrapper sx={{ mt: [2, '48px'], gap: ['88px', 3, 3] }}>
     {children}
   </ProductCardsWrapper>
 )
